+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The next release will require at least [Go 1.26].
 
 ### Added
 
+- Add `go.opentelemetry.io/contrib/detectors/oraclecloud`, a new resource detector for Oracle Cloud Infrastructure (OCI) instances, ported from `processor/resourcedetectionprocessor/internal/oraclecloud` in `opentelemetry-collector-contrib`. Detects `cloud.provider`, `cloud.platform`, `cloud.region`, `cloud.availability_zone`, `host.id`, `host.name`, `host.type`, `k8s.cluster.name`, and `oracle_cloud.realm`. (#9407)
 - Add support for the `aws.ec2` resource detector in `go.opentelemetry.io/contrib/otelconf/x`. (#9139)
 - Support testing of [Go 1.27]. (#9524)